Copy Response fields into PgException, revert Response/ResponseImpl
