+++ vertx-pg-client/src/main/java/io/vertx/pgclient/PgException.java
@@ -18,7 +18,6 @@
 package io.vertx.pgclient;
 
 import io.vertx.core.json.Json;
-import io.vertx.pgclient.impl.codec.Response;
 
 /**
  * PostgreSQL error including all <a href="https://www.postgresql.org/docs/current/protocol-error-fields.html">fields
@@ -28,26 +27,64 @@ import io.vertx.pgclient.impl.codec.Response;
  */
 public class PgException extends RuntimeException {
 
-  private final Response response;
   private final String errorMessage;
   private final String severity;
   private final String code;
   private final String detail;
-
-  public PgException(Response response) {
-    this.response = response;
-    this.errorMessage = null;
-    this.severity = null;
-    this.code = null;
-    this.detail = null;
-  }
+  private final String hint;
+  private final String position;
+  private final String internalPosition;
+  private final String internalQuery;
+  private final String where;
+  private final String file;
+  private final String line;
+  private final String routine;
+  private final String schema;
+  private final String table;
+  private final String column;
+  private final String dataType;
+  private final String constraint;
 
   public PgException(String errorMessage, String severity, String code, String detail) {
-    this.response = null;
     this.errorMessage = errorMessage;
     this.severity = severity;
     this.code = code;
     this.detail = detail;
+    this.hint = null;
+    this.position = null;
+    this.internalPosition = null;
+    this.internalQuery = null;
+    this.where = null;
+    this.file = null;
+    this.line = null;
+    this.routine = null;
+    this.schema = null;
+    this.table = null;
+    this.column = null;
+    this.dataType = null;
+    this.constraint = null;
+  }
+
+  public PgException(String errorMessage, String severity, String code, String detail, String hint, String position,
+      String internalPosition, String internalQuery, String where, String file, String line, String routine,
+      String schema, String table, String column, String dataType, String constraint) {
+    this.errorMessage = errorMessage;
+    this.severity = severity;
+    this.code = code;
+    this.detail = detail;
+    this.hint = hint;
+    this.position = position;
+    this.internalPosition = internalPosition;
+    this.internalQuery = internalQuery;
+    this.where = where;
+    this.file = file;
+    this.line = line;
+    this.routine = routine;
+    this.schema = schema;
+    this.table = table;
+    this.column = column;
+    this.dataType = dataType;
+    this.constraint = constraint;
   }
 
   /**
@@ -56,7 +93,7 @@ public class PgException extends RuntimeException {
    */
   public String getErrorMessage() {
     // getErrorMessage() avoids name clash with RuntimeException#getMessage()
-    return response == null ? errorMessage : response.getMessage();
+    return errorMessage;
   }
 
   /**
@@ -64,7 +101,7 @@ public class PgException extends RuntimeException {
    *     (<a href="https://www.postgresql.org/docs/current/protocol-error-fields.html">'S' field</a>)
    */
   public String getSeverity() {
-    return response == null ? severity : response.getSeverity();
+    return severity;
   }
 
   /**
@@ -74,68 +111,126 @@ public class PgException extends RuntimeException {
    *     it is never localized
    */
   public String getCode() {
-    return response == null ? code : response.getCode();
+    return code;
   }
 
   /**
    * @return an optional secondary error message carrying more detail about the problem
-   *     (<a href="https://www.postgresql.org/docs/current/protocol-error-fields.html">'D' field</a>).
+   *     (<a href="https://www.postgresql.org/docs/current/protocol-error-fields.html">'D' field</a>),
    *     a newline indicates paragraph break.
    */
   public String getDetail() {
-    return response == null ? detail : response.getDetail();
+    return detail;
   }
 
+  /**
+   * @return an optional suggestion (advice) what to do about the problem
+   *     (<a href="https://www.postgresql.org/docs/current/protocol-error-fields.html">'H' field</a>),
+   *     a newline indicates paragraph break.
+   */
   public String getHint() {
-    return response == null ? null : response.getHint();
+    return hint;
   }
 
+  /**
+   * @return a decimal ASCII integer, indicating an error cursor position as an index into the original
+   *     query string. (<a href="https://www.postgresql.org/docs/current/protocol-error-fields.html">'P' field</a>).
+   *     The first character has index 1, and positions are measured in characters not bytes.
+   */
   public String getPosition() {
-    return response == null ? null : response.getPosition();
+    return position;
   }
 
+  /**
+   * @return an indication of the context in which the error occurred
+   *     (<a href="https://www.postgresql.org/docs/current/protocol-error-fields.html">'W' field</a>).
+   *     Presently this includes a call stack traceback of active procedural language functions and
+   *     internally-generated queries. The trace is one entry per line, most recent first.
+   */
   public String getWhere() {
-    return response == null ? null : response.getWhere();
+    return where;
   }
 
+  /**
+   * @return file name of the source-code location where the error was reported
+   *     (<a href="https://www.postgresql.org/docs/current/protocol-error-fields.html">'F' field</a>).
+   */
   public String getFile() {
-    return response == null ? null : response.getFile();
+    return file;
   }
 
+  /**
+   * @return line number of the source-code location where the error was reported
+   *     (<a href="https://www.postgresql.org/docs/current/protocol-error-fields.html">'L' field</a>).
+   */
   public String getLine() {
-    return response == null ? null : response.getLine();
+    return line;
   }
 
+  /**
+   * @return name of the source-code routine reporting the error
+   *     (<a href="https://www.postgresql.org/docs/current/protocol-error-fields.html">'R' field</a>).
+   */
   public String getRoutine() {
-    return response == null ? null : response.getRoutine();
+    return routine;
   }
 
+  /**
+   * @return if the error was associated with a specific database object, the name of the schema containing
+   *     that object, if any
+   *     (<a href="https://www.postgresql.org/docs/current/protocol-error-fields.html">'s' field</a>).
+   */
   public String getSchema() {
-    return response == null ? null : response.getSchema();
+    return schema;
   }
 
+  /**
+   * @return if the error was associated with a specific table, the name of the table
+   *     (<a href="https://www.postgresql.org/docs/current/protocol-error-fields.html">'t' field</a>).
+   */
   public String getTable() {
-    return response == null ? null : response.getTable();
+    return table;
   }
 
+  /**
+   * @return if the error was associated with a specific table column, the name of the column
+   *     (<a href="https://www.postgresql.org/docs/current/protocol-error-fields.html">'c' field</a>).
+   */
   public String getColumn() {
-    return response == null ? null : response.getColumn();
+    return column;
   }
 
+  /**
+   * @return if the error was associated with a specific data type, the name of the data type
+   *     (<a href="https://www.postgresql.org/docs/current/protocol-error-fields.html">'d' field</a>).
+   */
   public String getDataType() {
-    return response == null ? null : response.getDataType();
+    return dataType;
   }
 
+  /**
+   * @return if the error was associated with a specific constraint, the name of the constraint
+   *     (<a href="https://www.postgresql.org/docs/current/protocol-error-fields.html">'n' field</a>).
+   */
   public String getConstraint() {
-    return response == null ? null : response.getConstraint();
+    return constraint;
   }
 
+  /**
+   * @return a decimal ASCII integer, indicating an error cursor position
+   *     (<a href="https://www.postgresql.org/docs/current/protocol-error-fields.html">'p' field</a>)
+   *     as an index into the internally generated command (see 'q' field).
+   */
   public String getInternalPosition() {
-    return response == null ? null : response.getInternalPosition();
+    return internalPosition;
   }
 
+  /**
+   * @return the text of a failed internally-generated command
+   *     (<a href="https://www.postgresql.org/docs/current/protocol-error-fields.html">'q' field</a>).
+   */
   public String getInternalQuery() {
-    return response == null ? null : response.getInternalQuery();
+    return internalQuery;
   }
 
   private static void append(StringBuffer stringBuffer, String key, String value) {
