+++ vertx-pg-client/src/main/java/io/vertx/pgclient/impl/codec/NoticeResponse.java
@@ -24,7 +24,7 @@ import io.vertx.sqlclient.impl.Notice;
  * @author <a href="mailto:emad.albloushi@gmail.com">Emad Alblueshi</a>
  */
 
-class NoticeResponse extends ResponseImpl implements Notice {
+class NoticeResponse extends Response implements Notice {
 
   @Override
   public void log(Logger logger) {
