+++ vertx-pg-client/src/main/java/io/vertx/pgclient/impl/codec/PgDecoder.java
@@ -239,7 +239,7 @@ class PgDecoder extends ChannelInboundHandlerAdapter {
     inflight.peek().handleNoticeResponse(response);
   }
 
-  private void decodeErrorOrNotice(ResponseImpl response, ByteBuf in) {
+  private void decodeErrorOrNotice(Response response, ByteBuf in) {
 
     byte type;
